Use IMDSv2 metadata service
Since 2019[1], EC2 has provided a more secure protocol for interacting
with the local metadata service. This protocol requires fetching a
token via a PUT request, and then using that token in an HTTP header
for subsequent GET requests to the metadata API. This adds significant
protection from a variety of SSRF attacks which could expose instance
profile credentials or other internal metadata.

This commit enables requiring the HTTP token via the EC2 Launch
Template resouce, and updates the install_vault.sh script to fetch
and use the IMDSv2 token for its metadata requests.

[1] https://aws.amazon.com/blogs/security/defense-in-depth-open-firewalls-reverse-proxies-ssrf-vulnerabilities-ec2-instance-metadata-service/